Anomaly Detection
Here I have shown how the Amount column is been affected on the basis of Mismatched of Fuel_Type values i.e. DIESEL and PETROL and this mismatched of values is been shown on the basis of graphs with PTROCRD column as PTROCRD column just tell us that each employee is been given a particular Card so whenever they will go to the PETROL PUMP then they could show this card to the Man present at the PETROL PUMP and could easily pay the amount for the Fuel.