Anomaly Detection
Here I have shown how the Resultant variable i.e. Amount column is been affected due to mismatching of Values present on the basis of Fuel_Price Anomaly i.e. here we first merged the two datasets then we generally used to see the mismatched of Fuel_Capacity values know on the basis of that we calculate I litre Fuel Price and then we finally calculate Original_Amount and then we compare the Analysis_Amount and Original_Amount and calculate Extra_Amount by just subtracting the Original_Amount from the Analysis_Amount and then we show this analysis on the basis of graph.