Anomaly Detection
Actually this time I m posting the all new project of mine and this time the datasets which I got is related to Marketing and Sales data i.e. here I m dealing with the datasets which contain the data regarding Sales and my job is do analysis on the basis of overall sales of the products for each outlet i.e. shop and shop which the datasets contain is generally related to supermarkets and grocery shops so this time I m showing the concept of Anomaly on the basis of Fat_Content i.e. Here I m analyzing how due to mismatched of values present in the Fat_Content column the overall Price of the products get increased in the analysis dataset when compared to original dataset and then we use to calculate the Original_Sales and try to compare it with Analysis_Sales and then see how much Extra_Sales is been earned by a particular Shopkeeper owner and this I have to do for each product i.e. here I will see the overall sales for each product on the basis of each Outlet.